+++ firmware/test/stub/tl_common.h
@@ -4,6 +4,7 @@
 #include <string.h>
 
 typedef enum { AS_GPIO = 0 } GPIO_FuncTypeDef;
+typedef uint32_t u32;
 typedef int bool;
 #define true 1
 #define false 0
@@ -13,3 +14,5 @@ void gpio_set_output_en(uint32_t pin, unsigned value);
 void gpio_set_input_en(uint32_t pin, unsigned value);
 void gpio_set_func(uint32_t pin, GPIO_FuncTypeDef func);
 void sleep_us(unsigned long us);
+u32 drv_disable_irq(void);
+u32 drv_restore_irq(u32 en);